+++ sdk/monitor/Azure.Monitor.OpenTelemetry.LiveMetrics/src/Internals/DataCollection/DoubleBuffer.cs
@@ -4,7 +4,7 @@
 using System.Threading;
 using Azure.Monitor.OpenTelemetry.LiveMetrics.Models;
 
-namespace Azure.Monitor.OpenTelemetry.LiveMetrics.Internals
+namespace Azure.Monitor.OpenTelemetry.LiveMetrics.Internals.DataCollection
 {
     /// <summary>
     /// Implements a double buffering mechanism for handling DocumentIngress objects.
